Added dedicated script for sensors (sensor.py) including a PositionSensor
