+++ tests/integration_tests/test_multiple_sensors.py
@@ -18,7 +18,7 @@ import numpy as np
 
 # local imports
 from probeye.definition.forward_model import ForwardModelTemplate
-from probeye.definition.forward_model import Sensor
+from probeye.definition.sensor import Sensor, PositionSensor
 from probeye.definition.inference_problem import InferenceProblem
 from probeye.inference.taralli_.noise_models import NormalNoise
 from probeye.inference.taralli_.solver import run_taralli_solver
@@ -85,11 +85,6 @@ class TestProblem(unittest.TestCase):
         #                       Define the Forward Model                       #
         # ==================================================================== #
 
-        class PositionSensor(Sensor):
-            def __init__(self, name, position):
-                super().__init__(name)
-                self.position = position
-
         class LinearModel(ForwardModelTemplate):
             def __call__(self, inp):
                 t = inp['time']
@@ -97,7 +92,7 @@ class TestProblem(unittest.TestCase):
                 B = inp['B']
                 response = dict()
                 for os in self.output_sensors:
-                    response[os.name] = A * os.position + B * t
+                    response[os.name] = A * os.x + B * t
                 return response
 
         # ==================================================================== #
@@ -136,9 +131,9 @@ class TestProblem(unittest.TestCase):
 
         # add the forward model to the problem
         isensor = Sensor("time")
-        osensor1 = PositionSensor("y1", pos_s1)
-        osensor2 = PositionSensor("y2", pos_s2)
-        osensor3 = PositionSensor("y3", pos_s3)
+        osensor1 = PositionSensor("y1", x=pos_s1)
+        osensor2 = PositionSensor("y2", x=pos_s2)
+        osensor3 = PositionSensor("y3", x=pos_s3)
         linear_model = LinearModel(
             ['A', 'B'], [isensor], [osensor1, osensor2, osensor3])
         problem.add_forward_model("LinearModel", linear_model)
